fix: move CertificateTile to a separate file
